fix: remove .has method from interface
The `.has` method is a trivial way of introducing race conditions into
calling code as since it's async there's nothing to stop the value for
the passed key being removed or added to the datastore before the next
line of code is executed.

Instead the caller should just call `.get` for the key and then handle
the `notFoundError` that will be thrown if there is no value present
for the key.

A common use for this method is to not put values into the datastore
that already exist - datastore implementations should perform this
optimisation, not calling code.

BREAKING CHANGE:

- The `.has` method has been removed, call `.get` instead