test: add tests for mutating the datastore during a query
When doing things like repo migrations we need to be able to mutate
the datastore while iterating over a list of things in the datastore.

Similarly long running processes can query the datastore while it's
being mutated by add operations, etc.

There are no tests for this, so here we add a that mutates the
datastore during a query and one that runs a query while the datastore
is being mutated.